+++ src/SoftTissue.Core/ConstitutiveEquations/QuasiLinearModel/Fung/FungModel.cs
@@ -7,15 +7,16 @@ using System;
 
 namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel.Fung
 {
+    /// <summary>
+    /// It represents the viscoelastic Fung Model.
+    /// </summary>
     public class FungModel : QuasiLinearViscoelasticityModel<FungModelInput, FungModelResult, ReducedRelaxationFunctionData>, IFungModel
     {
         /// <summary>
         /// Class constructor.
         /// </summary>
         /// <param name="simpsonRuleIntegration"></param>
-        public FungModel(ISimpsonRuleIntegration simpsonRuleIntegration, IDerivative derivative)
-            : base(simpsonRuleIntegration, derivative)
-        { }
+        public FungModel(ISimpsonRuleIntegration simpsonRuleIntegration, IDerivative derivative) : base(simpsonRuleIntegration, derivative) { }
 
         /// <summary>
         /// The time when the alternative and original reduced relaxation function converge.
@@ -31,14 +32,14 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel.Fung
         public override double CalculateReducedRelaxationFunction(FungModelInput input, double time)
         {
             // When considering that the viscoelastic effect ocurrer just after the ramp time, the reduced relaxation function must not
-            // be considered in calculations and, after the rampa time, the time that will be used, must be adjusted to the initial time
+            // be considered in calculations and, after the ramp time, the time that will be used, must be adjusted to the initial time
             // be the ramp time.
             if (input.ViscoelasticConsideration == ViscoelasticConsideration.ViscoelasticEffectAfterRampTime)
             {
-                if (time <= input.RampTime)
+                if (time <= input.FirstRampTime)
                     return 1;
                 else
-                    time -= input.RampTime;
+                    time -= input.FirstRampTime;
             }
 
             // Here is applied the boundary conditions for Reduced Relaxation Function for time equals to zero.
@@ -48,12 +49,14 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel.Fung
             {
                 return 1;
             }
-            
+
             var reducedRelaxationFunctionInput = input.ReducedRelaxationFunctionInput;
 
             // The original equation was rewritten to simplify the E1 equation.
             // Instead of calculate E1 twice, was unified the dominion of both integration. See the explanation on Equation.docx.
-            return (1 + reducedRelaxationFunctionInput.RelaxationIndex * this.CalculateI(reducedRelaxationFunctionInput.SlowRelaxationTime, reducedRelaxationFunctionInput.FastRelaxationTime, input.TimeStep, time)) / (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime));
+            return
+                (1 + reducedRelaxationFunctionInput.RelaxationIndex * this.CalculateI(reducedRelaxationFunctionInput.SlowRelaxationTime, reducedRelaxationFunctionInput.FastRelaxationTime, input.TimeStep, time))
+                / (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime));
         }
 
         /// <summary>
@@ -71,12 +74,15 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel.Fung
             // be the ramp time.
             if (input.ViscoelasticConsideration == ViscoelasticConsideration.ViscoelasticEffectAfterRampTime)
             {
-                if (time <= input.RampTime)
+                if (time <= input.FirstRampTime)
                 {
-                    return reducedRelaxationFunctionInput.RelaxationIndex * (-1 / reducedRelaxationFunctionInput.FastRelaxationTime + 1 / reducedRelaxationFunctionInput.SlowRelaxationTime) / (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime));
+                    return
+                        reducedRelaxationFunctionInput.RelaxationIndex
+                        * (-1 / reducedRelaxationFunctionInput.FastRelaxationTime + 1 / reducedRelaxationFunctionInput.SlowRelaxationTime)
+                        / (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime));
                 }
                 else
-                    time -= input.RampTime;
+                    time -= input.FirstRampTime;
             }
 
             // Here is applied the boundary conditions for Reduced Relaxation Function for time equals to zero.
@@ -84,10 +90,16 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel.Fung
             // is called in another methods, that error must be considered to times near to zero.
             if (time <= Constants.Precision)
             {
-                return reducedRelaxationFunctionInput.RelaxationIndex * (-1 / reducedRelaxationFunctionInput.FastRelaxationTime + 1 / reducedRelaxationFunctionInput.SlowRelaxationTime) / (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime));
+                return
+                    reducedRelaxationFunctionInput.RelaxationIndex
+                    * (-1 / reducedRelaxationFunctionInput.FastRelaxationTime + 1 / reducedRelaxationFunctionInput.SlowRelaxationTime)
+                    / (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime));
             }
 
-            return (reducedRelaxationFunctionInput.RelaxationIndex * (Math.Exp(-time / reducedRelaxationFunctionInput.FastRelaxationTime) - Math.Exp(-time / reducedRelaxationFunctionInput.SlowRelaxationTime))) / (time * (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime)));
+            return
+                reducedRelaxationFunctionInput.RelaxationIndex
+                * (Math.Exp(-time / reducedRelaxationFunctionInput.FastRelaxationTime) - Math.Exp(-time / reducedRelaxationFunctionInput.SlowRelaxationTime))
+                / (time * (1 + reducedRelaxationFunctionInput.RelaxationIndex * Math.Log(reducedRelaxationFunctionInput.SlowRelaxationTime / reducedRelaxationFunctionInput.FastRelaxationTime)));
         }
 
         /// <summary>