+++ src/SoftTissue.Core/ConstitutiveEquations/QuasiLinearModel/Fung/IFungModel.cs
@@ -3,6 +3,9 @@ using SoftTissue.Infrastructure.Models;
 
 namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel.Fung
 {
+    /// <summary>
+    /// It represents the viscoelastic Fung Model.
+    /// </summary>
     public interface IFungModel : IQuasiLinearViscoelasticityModel<FungModelInput, FungModelResult, ReducedRelaxationFunctionData>
     {
         /// <summary>