+++ src/SoftTissue.Core/ConstitutiveEquations/QuasiLinearModel/IQuasiLinearViscoelasticityModel.cs
@@ -1,7 +1,14 @@
 ﻿using SoftTissue.Core.Models.Viscoelasticity.QuasiLinear;
+using System.Threading.Tasks;
 
 namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
 {
+    /// <summary>
+    /// It represents the quasi-linear viscoelastic model.
+    /// </summary>
+    /// <typeparam name="TInput"></typeparam>
+    /// <typeparam name="TResult"></typeparam>
+    /// <typeparam name="TRelaxationFunction"></typeparam>
     public interface IQuasiLinearViscoelasticityModel<TInput, TResult, TRelaxationFunction> : IViscoelasticModel<TInput>
         where TInput : QuasiLinearViscoelasticityModelInput<TRelaxationFunction>, new()
         where TResult : QuasiLinearViscoelasticityModelResult, new()
@@ -9,9 +16,16 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
         /// <summary>
         /// This method calculates the initial conditions for Fung model analysis.
         /// </summary>
-        /// <param name="input"></param>
         /// <returns></returns>
-        TResult CalculateInitialConditions(TInput input);
+        Task<TResult> CalculateInitialConditions();
+
+        /// <summary>
+        /// This method calculates the results for a quasi-linear viscoelastic model.
+        /// </summary>
+        /// <param name="input"></param>
+        /// <param name="time"></param>
+        /// <returns></returns>
+        Task<TResult> CalculateResults(TInput input, double time);
 
         /// <summary>
         /// This method calculates the strain derivative.