+++ src/SoftTissue.Core/ConstitutiveEquations/QuasiLinearModel/QuasiLinearViscoelasticityModel.cs
@@ -4,15 +4,28 @@ using SoftTissue.Core.NumericalMethods.Derivative;
 using SoftTissue.Core.NumericalMethods.Integral.Simpson;
 using SoftTissue.Infrastructure.Models;
 using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
 {
+    /// <summary>
+    /// It represents the quasi-linear viscoelastic model.
+    /// </summary>
+    /// <typeparam name="TInput"></typeparam>
+    /// <typeparam name="TResult"></typeparam>
+    /// <typeparam name="TRelaxationFunction"></typeparam>
     public abstract class QuasiLinearViscoelasticityModel<TInput, TResult, TRelaxationFunction> : ViscoelasticModel<TInput>, IQuasiLinearViscoelasticityModel<TInput, TResult, TRelaxationFunction>
         where TInput : QuasiLinearViscoelasticityModelInput<TRelaxationFunction>, new()
         where TResult : QuasiLinearViscoelasticityModelResult, new()
     {
-        private readonly ISimpsonRuleIntegration _simpsonRuleIntegration;
-        private readonly IDerivative _derivative;
+        /// <summary>
+        /// The important relaxation times.
+        /// </summary>
+        protected RelaxationTimes _relaxationTimes;
+
+        protected readonly ISimpsonRuleIntegration _simpsonRuleIntegration;
+        protected readonly IDerivative _derivative;
 
         /// <summary>
         /// Class constructor.
@@ -27,19 +40,102 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
         /// <summary>
         /// This method calculates the initial conditions for Fung model analysis.
         /// </summary>
-        /// <param name="input"></param>
         /// <returns></returns>
-        public virtual TResult CalculateInitialConditions(TInput input)
+        public virtual Task<TResult> CalculateInitialConditions()
         {
-            return new TResult
+            return Task.FromResult(new TResult
             {
                 ReducedRelaxationFunction = 1,
-                ElasticResponse = 0,
                 Strain = 0,
+                ElasticResponse = 0,
                 Stress = 0,
                 StressByIntegralDerivative = 0,
                 StressByReducedRelaxationFunctionDerivative = 0
+            });
+        }
+
+        /// <summary>
+        /// This method calculates the results for a quasi-linear viscoelastic model.
+        /// </summary>
+        /// <param name="input"></param>
+        /// <param name="time"></param>
+        /// <param name="streamWriter"></param>
+        public virtual async Task<TResult> CalculateResults(TInput input, double time)
+        {
+            if (time <= Constants.Precision)
+                return await this.CalculateInitialConditions();
+
+            input.RelaxationNumber = this.CalculateRelaxationNumber(input, time);
+
+            var tasks = new List<Task>();
+
+            double strain = 0;
+            tasks.Add(Task.Run(() =>
+            {
+                strain = this.CalculateStrain(input, time);
+            }));
+
+            double reducedRelaxationFunction = 0;
+            tasks.Add(Task.Run(() =>
+            {
+                reducedRelaxationFunction = this.CalculateReducedRelaxationFunction(input, time);
+            }));
+
+            double elasticResponse = 0;
+            tasks.Add(Task.Run(() =>
+            {
+                elasticResponse = this.CalculateElasticResponse(input, time);
+            }));
+
+            double stress = 0;
+            tasks.Add(Task.Run(() =>
+            {
+                stress = this.CalculateStress(input, time);
+            }));
+
+            double stressByReducedRelaxationFunctionDerivative = 0;
+            tasks.Add(Task.Run(() =>
+            {
+                stressByReducedRelaxationFunctionDerivative = this.CalculateStressByReducedRelaxationFunctionDerivative(input, time);
+            }));
+
+            double stressByIntegralDerivative = 0;
+            tasks.Add(Task.Run(() =>
+            {
+                stressByIntegralDerivative = this.CalculateStressByIntegralDerivative(input, time);
+            }));
+
+            await Task.WhenAll(tasks);
+
+            return new TResult
+            {
+                Strain = strain,
+                ReducedRelaxationFunction = reducedRelaxationFunction,
+                ElasticResponse = elasticResponse,
+                Stress = stress,
+                StressByReducedRelaxationFunctionDerivative = stressByReducedRelaxationFunctionDerivative,
+                StressByIntegralDerivative = stressByIntegralDerivative
             };
+        }
+
+        public int CalculateRelaxationNumber(TInput input, double time)
+        {
+            // If the time is less than the first relaxation total time, it is in the first relaxation.
+            if (time < input.FirstRelaxationTotalTime)
+                return 0;
+
+            for (int i = 1; i < input.NumerOfRelaxations; i++)
+            {
+                if (input.FirstRelaxationTotalTime + (i - 1) * input.RelaxationTotalTime < time &&
+                    time <= input.FirstRelaxationTotalTime + i * input.RelaxationTotalTime)
+                    return i;
+            }
+
+            // If the time is bigger than all relaxations time, it means that it is on the last relaxation.
+            if (time > input.FirstRelaxationTotalTime + (input.NumerOfRelaxations - 1) * input.RelaxationTotalTime)
+                return input.NumerOfRelaxations - 1;
+
+            throw new ArgumentOutOfRangeException("Time", $"Time: {time} is not in a valid range to calculate the relaxation number.");
         }
 
         /// <summary>
@@ -50,30 +146,53 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
         /// <returns></returns>
         public override double CalculateStrain(TInput input, double time)
         {
-            switch (input.ViscoelasticConsideration)
-            {
-                case ViscoelasticConsideration.GeneralViscoelasctiEffect:
-                case ViscoelasticConsideration.ViscoelasticEffectAfterRampTime:
-                    return time <= input.RampTime ? input.StrainRate * time : input.MaximumStrain;
+            if (this._relaxationTimes.Equals(default) == true)
+                this._relaxationTimes = this.BuildRelaxationTimes(input);
 
-                case ViscoelasticConsideration.GeneralViscoelasticEffectWithStrainDecrease:
-                case ViscoelasticConsideration.ViscoelasticEffectAfterRampTimeWithStrainDecrease:
-                    if (time <= input.RampTime)
-                        return input.StrainRate * time;
+            if (input.ViscoelasticConsideration == ViscoelasticConsideration.DisregardRampTime)
+                return input.MaximumStrain;
 
-                    if (time > input.RampTime && time <= input.TimeWhenStrainStartDecreasing)
-                        return input.MaximumStrain;
+            // time = 0 
+            //     --> strain = 0
+            if (time <= Constants.Precision)
+                return 0;
 
-                    if (time > input.TimeWhenStrainStartDecreasing && time <= input.TimeWhenStrainStartDecreasing + input.DecreaseTime)
-                        return input.MaximumStrain - input.StrainDecreaseRate * (time - input.TimeWhenStrainStartDecreasing);
+            // 0 < time <= first ramp time
+            //     --> strain = strain rate * time
+            if (time > Constants.Precision && time <= input.FirstRampTime)
+                return input.StrainRate * time;
 
-                    return input.MinimumStrain;
+            // first ramp time < time <= first relaxation total time
+            // OBS: first relaxation total time = first ramp time + time with constant maximum strain
+            //     --> strain = maximum strain
+            if (time > input.FirstRampTime && time <= input.FirstRelaxationTotalTime)
+                return input.MaximumStrain;
 
-                case ViscoelasticConsideration.DisregardRampTime:
-                    return input.MaximumStrain;
-            }
+            // This part is used only for the second relaxations and next.
+            if (time > input.FirstRelaxationTotalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainDecreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return input.MaximumStrain -
+                    input.StrainDecreaseRate * (time - (input.FirstRelaxationTotalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime));
 
-            throw new ArgumentOutOfRangeException("Invalid viscoelastic consideration.");
+            if (time > this._relaxationTimes.StrainDecreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainDecreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return input.MinimumStrain;
+
+            if (time > this._relaxationTimes.StrainDecreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainIncreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return input.MinimumStrain +
+                    input.StrainRate * (time - (this._relaxationTimes.StrainDecreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime));
+
+            if (time > this._relaxationTimes.StrainIncreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainIncreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return input.MaximumStrain;
+
+            // If the time is bigger than all relaxations time, it means that it is on the last relaxation and the strain is kept at the maximum till the end of analysis.
+            if (time > input.FirstRelaxationTotalTime + (input.NumerOfRelaxations - 1) * input.RelaxationTotalTime)
+                return input.MaximumStrain;
+
+            // The default value returned must be equals to zero.
+            return 0;
         }
 
         /// <summary>
@@ -84,34 +203,36 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
         /// <returns></returns>
         public virtual double CalculateStrainDerivative(TInput input, double time)
         {
-            switch (input.ViscoelasticConsideration)
-            {
-                case ViscoelasticConsideration.GeneralViscoelasctiEffect:
-                case ViscoelasticConsideration.ViscoelasticEffectAfterRampTime:
-                    return time < input.RampTime ? input.StrainRate : 0;
+            if (this._relaxationTimes.Equals(default) == true)
+                this._relaxationTimes = this.BuildRelaxationTimes(input);
 
-                case ViscoelasticConsideration.GeneralViscoelasticEffectWithStrainDecrease:
-                case ViscoelasticConsideration.ViscoelasticEffectAfterRampTimeWithStrainDecrease:
-                    {
-                        if (time <= input.RampTime)
-                            return input.StrainRate;
+            if (input.ViscoelasticConsideration == ViscoelasticConsideration.DisregardRampTime)
+                return 0;
 
-                        double timeWhenStrainBeginningDecrease = input.RampTime + input.TimeWithConstantStrain;
+            if (time <= input.FirstRampTime)
+                return input.StrainRate;
 
-                        if (time > input.RampTime && time <= timeWhenStrainBeginningDecrease)
-                            return 0;
+            if (time > input.FirstRampTime && time <= input.FirstRelaxationTotalTime)
+                return 0;
 
-                        if (time > timeWhenStrainBeginningDecrease && time <= timeWhenStrainBeginningDecrease + input.DecreaseTime)
-                            return -input.StrainDecreaseRate;
+            if (time > input.FirstRelaxationTotalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainDecreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return -input.StrainDecreaseRate;
 
-                        return 0;
-                    }
+            if (time > this._relaxationTimes.StrainDecreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainDecreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return 0;
 
-                case ViscoelasticConsideration.DisregardRampTime:
-                    return 0;
-            }
+            if (time > this._relaxationTimes.StrainDecreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainIncreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return input.StrainRate;
 
-            throw new ArgumentOutOfRangeException("Invalid viscoelastic consideration.");
+            if (time > this._relaxationTimes.StrainIncreaseStartTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime &&
+                time <= this._relaxationTimes.StrainIncreaseFinalTime + (input.RelaxationNumber - 1) * input.RelaxationTotalTime)
+                return 0;
+
+            // The default value returned must be equals to zero.
+            return 0;
         }
 
         /// <summary>
@@ -125,9 +246,8 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
             if (input.ViscoelasticConsideration == ViscoelasticConsideration.DisregardRampTime)
                 return input.InitialStress;
 
-            double strain = this.CalculateStrain(input, time);
-
-            return input.ElasticStressConstant * (Math.Exp(input.ElasticPowerConstant * strain) - 1);
+            // Elastic stress = A * [exp(B * strain) - 1]
+            return input.ElasticStressConstant * (Math.Exp(input.ElasticPowerConstant * this.CalculateStrain(input, time)) - 1);
         }
 
         /// <summary>
@@ -144,6 +264,7 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
             double strain = this.CalculateStrain(input, time);
             double strainDerivative = this.CalculateStrainDerivative(input, time);
 
+            // Derivative of elastic stress = A * B * (d/dt)(strain) * exp(B * strain)
             return input.ElasticStressConstant * input.ElasticPowerConstant * strainDerivative * Math.Exp(input.ElasticPowerConstant * strain);
         }
 
@@ -177,7 +298,7 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
             if (time <= Constants.Precision)
                 return 0;
 
-            if (time <= input.RampTime && time > Constants.Precision)
+            if (time > Constants.Precision && time <= input.FirstRampTime)
             {
                 if (input.ViscoelasticConsideration == ViscoelasticConsideration.ViscoelasticEffectAfterRampTime)
                     return this.CalculateElasticResponse(input, time);
@@ -192,37 +313,37 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
                     });
             }
 
-            if (time <= input.TimeWhenStrainStartDecreasing && time > input.RampTime)
+            if (time <= input.FirstRelaxationTotalTime && time > input.FirstRampTime)
                 return this._simpsonRuleIntegration.Calculate(
                     (integrationTime) => this.CalculateReducedRelaxationFunction(input, time - integrationTime) * this.CalculateElasticResponseDerivative(input, integrationTime),
                     new IntegralInput
                     {
                         InitialPoint = 0,
-                        FinalPoint = input.RampTime,
+                        FinalPoint = input.FirstRampTime,
                         Step = input.TimeStep
                     });
 
-            if (input.ViscoelasticConsideration == ViscoelasticConsideration.GeneralViscoelasticEffectWithStrainDecrease
-                || input.ViscoelasticConsideration == ViscoelasticConsideration.ViscoelasticEffectAfterRampTimeWithStrainDecrease)
-            {
-                return
-                    this._simpsonRuleIntegration.Calculate(
-                        (integrationTime) => this.CalculateReducedRelaxationFunction(input, time - integrationTime) * this.CalculateElasticResponseDerivative(input, integrationTime),
-                        new IntegralInput
-                        {
-                            InitialPoint = 0,
-                            FinalPoint = input.RampTime,
-                            Step = input.TimeStep
-                        })
-                    + this._simpsonRuleIntegration.Calculate(
-                        (integrationTime) => this.CalculateReducedRelaxationFunction(input, time - integrationTime) * this.CalculateElasticResponseDerivative(input, integrationTime),
-                        new IntegralInput
-                        {
-                            InitialPoint = input.TimeWhenStrainStartDecreasing,
-                            FinalPoint = time <= input.TimeWhenStrainStartDecreasing + input.DecreaseTime ? time : input.TimeWhenStrainStartDecreasing + input.DecreaseTime,
-                            Step = input.TimeStep
-                        });
-            }
+            //if (input.ViscoelasticConsideration == ViscoelasticConsideration.GeneralViscoelasticEffectWithStrainDecrease
+            //    || input.ViscoelasticConsideration == ViscoelasticConsideration.ViscoelasticEffectAfterRampTimeWithStrainDecrease)
+            //{
+            //    return
+            //        this._simpsonRuleIntegration.Calculate(
+            //            (integrationTime) => this.CalculateReducedRelaxationFunction(input, time - integrationTime) * this.CalculateElasticResponseDerivative(input, integrationTime),
+            //            new IntegralInput
+            //            {
+            //                InitialPoint = 0,
+            //                FinalPoint = input.FirstRampTime,
+            //                Step = input.TimeStep
+            //            })
+            //        + this._simpsonRuleIntegration.Calculate(
+            //            (integrationTime) => this.CalculateReducedRelaxationFunction(input, time - integrationTime) * this.CalculateElasticResponseDerivative(input, integrationTime),
+            //            new IntegralInput
+            //            {
+            //                InitialPoint = input.FirstRelaxationTotalTime,
+            //                FinalPoint = time <= input.FirstRelaxationTotalTime + input.DecreaseTime ? time : input.FirstRelaxationTotalTime + input.DecreaseTime,
+            //                Step = input.TimeStep
+            //            });
+            //}
 
             // The default way to calculate the stress is using the convolution between the reduced relaxation function and derivative of elastic response.
             return this._simpsonRuleIntegration.Calculate(
@@ -281,5 +402,19 @@ namespace SoftTissue.Core.ConstitutiveEquations.QuasiLinearModel
                     }),
                 input.TimeStep, time);
         }
+
+        /// <summary>
+        /// This method calculates the important relaxation times.
+        /// </summary>
+        /// <param name="input"></param>
+        /// <returns></returns>
+        protected RelaxationTimes BuildRelaxationTimes(TInput input)
+            => new RelaxationTimes
+            (
+                input.FirstRelaxationTotalTime + input.DecreaseTime,
+                input.FirstRelaxationTotalTime + input.DecreaseTime + input.TimeWithConstantMinimumStrain,
+                input.FirstRelaxationTotalTime + input.DecreaseTime + input.TimeWithConstantMinimumStrain + input.RampTime,
+                input.FirstRelaxationTotalTime + input.DecreaseTime + input.TimeWithConstantMinimumStrain + input.RampTime + input.TimeWithConstantMaximumStrain
+            );
     }
 }