+++ src/SoftTissue.Core/Operations/QuasiLinearViscoelasticity/CalculateStress/CalculateQuasiLinearViscoelasticityStress.cs
@@ -35,76 +35,14 @@ namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress
 
         /// <summary>
         /// The header to solution file.
-        /// This property depends on what is calculated on method <see cref="CalculateAndWriteResults"/>.
         /// </summary>
-        protected override string SolutionFileHeader => $"Time;Strain;Reduced Relaxation Function;Elastic Response;Stress;Stress by dG;Stress by Integral Derivative";
+        protected override string SolutionFileHeader => $"Time,Strain,Reduced Relaxation Function,Elastic Response,Stress,Stress by dG,Stress by Integral Derivative";
 
         /// <summary>
         /// This method writes the input data into a file.
         /// </summary>
         /// <param name="input"></param>
         public abstract void WriteInput(TInput input);
-
-        /// <summary>
-        /// This method calculates the results and writes them to a file.
-        /// </summary>
-        /// <param name="input"></param>
-        /// <param name="time"></param>
-        /// <param name="streamWriter"></param>
-        public virtual async Task<TResult> CalculateAndWriteResults(TInput input, double time, StreamWriter streamWriter)
-        {
-            var tasks = new List<Task>();
-
-            double strain = 0;
-            tasks.Add(Task.Run(() =>
-            {
-                strain = this._viscoelasticModel.CalculateStrain(input, time);
-            }));
-
-            double reducedRelaxationFunction = 0;
-            tasks.Add(Task.Run(() =>
-            {
-                reducedRelaxationFunction = this._viscoelasticModel.CalculateReducedRelaxationFunction(input, time);
-            }));
-
-            double elasticResponse = 0;
-            tasks.Add(Task.Run(() =>
-            {
-                elasticResponse = this._viscoelasticModel.CalculateElasticResponse(input, time);
-            }));
-
-            double stress = 0;
-            tasks.Add(Task.Run(() =>
-            {
-                stress = this._viscoelasticModel.CalculateStress(input, time);
-            }));
-
-            double stressByReducedRelaxationFunctionDerivative = 0;
-            tasks.Add(Task.Run(() =>
-            {
-                stressByReducedRelaxationFunctionDerivative = this._viscoelasticModel.CalculateStressByReducedRelaxationFunctionDerivative(input, time);
-            }));
-
-            double stressByIntegralDerivative = 0;
-            tasks.Add(Task.Run(() =>
-            {
-                stressByIntegralDerivative = this._viscoelasticModel.CalculateStressByIntegralDerivative(input, time);
-            }));
-
-            await Task.WhenAll(tasks);
-
-            streamWriter.WriteLine($"{time};{strain};{reducedRelaxationFunction};{elasticResponse};{stress};{stressByReducedRelaxationFunctionDerivative};{stressByIntegralDerivative}");
-
-            return new TResult
-            {
-                Strain = strain,
-                ReducedRelaxationFunction = reducedRelaxationFunction,
-                ElasticResponse = elasticResponse,
-                Stress = stress,
-                StressByReducedRelaxationFunctionDerivative = stressByReducedRelaxationFunctionDerivative,
-                StressByIntegralDerivative = stressByIntegralDerivative
-            };
-        }
 
         /// <summary>
         /// This method executes an analysis to calculate the stress for a quasi-linear viscoelasticity model.
@@ -125,27 +63,30 @@ namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress
                 tasks.Add(Task.Run(async () =>
                 {
                     this.WriteInput(input);
-                    
+
                     double time = input.InitialTime;
-                    
-                    TResult previousResult = this._viscoelasticModel.CalculateInitialConditions(input);
-                    
+
+                    var previousResult = new TResult();
+
                     try
                     {
-                        using (StreamWriter streamWriter = new StreamWriter(this.CreateSolutionFile(input)))
+                        using (var streamWriter = new StreamWriter(this.CreateSolutionFile(input)))
                         {
                             streamWriter.WriteLine(this.SolutionFileHeader);
-                    
+
                             while (time <= input.FinalTime)
                             {
-                                TResult result = await this.CalculateAndWriteResults(input, time, streamWriter).ConfigureAwait(false);
-                    
+                                TResult result = await this._viscoelasticModel.CalculateResults(input, time).ConfigureAwait(false);
+
+                                streamWriter.WriteLine($"{time},{result}");
+
                                 // It increases the time step when the stress is converging to its asymptote.
-                                if (Math.Abs((result.Stress - previousResult.Stress) / previousResult.Stress) < Constants.RelativePrecision)
-                                    time += 10 * input.TimeStep;
-                                else
-                                    time += input.TimeStep;
-                    
+                                //if (Math.Abs((result.Stress - previousResult.Stress) / previousResult.Stress) < Constants.RelativePrecision)
+                                //    time += 10 * input.TimeStep;
+                                //else
+                                //    time += input.TimeStep;
+                                time += input.TimeStep;
+
                                 previousResult = result;
                             }
                         }
@@ -160,6 +101,11 @@ namespace SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress
             }
 
             return response;
+        }
+
+        protected int CalculateRelaxationNumber(TInput input, double time)
+        {
+            throw new NotImplementedException();
         }
     }
 }