+++ etc/CHANGELOG.md
@@ -14,7 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Operation to calculate stress for experimental models.
 - Enum ViscoelasticConsideration with each viscoelastic consideration.
 - Nuget package CsvHelper.
+- Nuget package Microsoft.AspNetCore.Mvc.NewtonsoftJson.
+- Nuget package Newtonsoft.Json.
 ### Changed
+- Quasi-linear operations considering multiple relaxations.
 - Data contracts of operations that inheriting from CalculteResult inherite from CalculateResult data contract.
 - RelavativePrecision from 5e-3 to 1e-3 and Precision from 1e-10 to 1e-6.
 - Moved sensitivity analysis operations to a new folder.